requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
pymongo==3.11.4
prometheus_client==0.11.0
python-dotenv==0.19.2
